# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JamJam Trek API",
        "description": "API documentation for JamJam Trek application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://161.97.167.73:8001",
            "description": "Server"
        }
    ],
    "paths": {
...
        "/api/blogs/{title}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get a single blog",
                "description": "Retrieve a blog by its title slug (lowercase with dashes)",
                "operationId": "getBlog",
                "parameters": [
                    {
                        "name": "title",
                        "in": "path",
                        "description": "Blog title slug (e.g., 'my-first-blog')",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Blog fetched successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "My First Blog"
                                                },
                                                "subtitle": {
                                                    "type": "string",
                                                    "example": "A subtitle"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Blog description"
                                                },
                                                "excerpt": {
                                                    "type": "string",
                                                    "example": "Short excerpt"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "my-first-blog"
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "blogs/image.jpg"
                                                },
                                                "image_url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/storage/blogs/image.jpg"
                                                },
                                                "content": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "heading": {
                                                                "type": "string"
                                                            },
                                                            "paragraph": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "conclusion": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Blog not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "/api/blogs/total": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get total blogs count",
                "description": "Retrieve the total number of blogs in the system",
                "operationId": "getTotalBlogs",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "total_blogs": {
                                            "type": "integer",
                                            "example": 42
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}